test(agent): move permission control examples into ignored tests
Replace prompt/streaming permission examples with ignored integration
tests and log tool-call rejections using structured fields.
